Updated cairo to 1.18.2.

+++ util/cairo-script/cairo-script-scanner.c
@@ -1559,7 +1559,7 @@ _translate_real (csi_t *ctx,
 {
     uint8_t hdr;
 
-    if (real >= INT32_MIN && real <= INT32_MAX && (int) real == real)
+    if ((double)real >= INT32_MIN && (double)real <= INT32_MAX && (int) real == real)
 	return _translate_integer (ctx, real, closure);
 
 #if WORDS_BIGENDIAN
